+++ src/hla/aptools.py
@@ -12,7 +12,10 @@ Accelerator Physics Tools
 import numpy as np
 import time, shelve
 import matplotlib.pylab as plt
+
+import machines
 from orbit import Orbit
+from catools import caput, caget 
 
 alphac = 3.6261976841792413e-04
 
@@ -176,18 +179,21 @@ def measDispersion():
 def correctOrbitPv(bpm, trim, ormdata):
     """
     correct orbit use direct pv and catools
+
+    - the input bpm and trim should be uniq in pv names.
     """
+    #print "Matrix size: ", len(bpm), len(trim)
     m = np.zeros((len(bpm), len(trim)), 'd')
-    for i,b in bpm:
+    for i,b in enumerate(bpm):
         im = ormdata.index(b)
         if im < 0: raise ValueError("PV %s is not found in ormdata" % b)
-        for j,t in trim:
+        for j,t in enumerate(trim):
             jm = ormdata.index(t)
             if jm < 0: raise ValueError("pv %s is not found in ormdata" % t)
             # did not check if the item is masked.
             m[i,j] = ormdata.m[im,jm]
 
-    v0 = caget(bpm)
+    v0 = np.array(caget(bpm), 'd')
     dk, resids, rank, s = np.linalg.lstsq(m, -1.0*v0)
     k0 = np.array(caget(trim), 'd')
     caput(trim, k0+dk)
@@ -200,53 +206,44 @@ def correctOrbit(bpm, trim, **kwargs):
 
       correctOrbit(['BPM1', 'BPM2'], ['T1', 'T2', 'T3'])
 
-    Since each bpm can have the 'X' or 'Y' reading, there is a parameter to
-    specify the plane:
-
-    - *plane* ['XX', 'XY', 'YY', 'YX']. The first 'X' or 'Y' is BPM and second
-       is Trim. If that BPM or Trim does not have readings from such plane,
-       this element should be ignored.
-    - *pv* use PV instead of element name
-    .. seealso:: `hla.getSubOrm`
+    .. seealso:: :func:`hla.getSubOrm`
     """
 
+    plane = kwargs.get('plane', 'HV')
+
     # an orbit based these bpm
-    obt = Orbit(bpm)
+    bpmlst = machines._lat.getElements(bpm)
+    pvx = [e.pv(tags=[machines.HLA_TAG_X, machines.HLA_TAG_EGET])
+           for e in bpmlst]
+    pvy = [e.pv(tags=[machines.HLA_TAG_Y, machines.HLA_TAG_EGET])
+           for e in bpmlst]
+
+    if plane == 'H': bpmpv = set(pvx)
+    elif plane == 'V': bpmpv = set(pvy)
+    else: bpmpv = set(pvx + pvy)
 
     # pv for trim
-    pvxrb = [e.pv(tags=[machines.HLA_TAG_X, machines.HLA_TAG_EGET]) 
-             for e in trim]
-    pvyrb = [e.pv(tags=[machines.HLA_TAG_Y, machines.HLA_TAG_EGET])
-             for e in trim]
-    pvxsp = [e.pv(tags=[machines.HLA_TAG_X, machines.HLA_TAG_EPUT]) 
-             for e in trim]
-    pvysp = [e.pv(tags=[machines.HLA_TAG_Y, machines.HLA_TAG_EPUT])
-             for e in trim]
+    trimlst = machines._lat.getElements(trim)
 
-    trimx = Element(**{'name':'HLA:ORBIT:X',
-                       'eget': [(trim[0].eget(), pvxrb, "x")],
-                       'eput': [(trim[0].eput(), pvxsp, "x")]})
-    trimy = Element(**{'name':'HLA:ORBIT:Y',
-                       'eget': [(trim[0].eget(), pvyrb, "y")],
-                       'eput': [(trim[0].eput(), pvysp, "y")]})
+    trimpv, pvxsp, pvysp = [], [], []
+    for e in trimlst:
+        pv = e.pv(tags=[machines.HLA_TAG_X, machines.HLA_TAG_EPUT])
+        if pv: pvxsp.append(pv)
+        if isinstance(pv, (str, unicode)): trimpv.append(pv)
+        elif pv: trimpv.extend(pv)
+        
+        pv = e.pv(tags=[machines.HLA_TAG_Y, machines.HLA_TAG_EPUT])
+        if pv: pvysp.append(pv)
+        if isinstance(pv, (str, unicode)): trimpv.append(pv)
+        else: trimpv.extend(pv)
 
-    plane = kwargs.get('plane', 'XX')
-    
-    m = getSubOrm([e.name for e in bpm], [e.name for e in trim], plane)
-    if plane[0] == 'X': v = getOrbit(bpm)[:,0]
-    elif plane[0] == 'Y': v = getOrbit(bpm)[:,1]
+    if 'H' in plane and len(pvx) > 0 and len(pvxsp) == 0:
+        print "WARNING: no HCOR for horizontal orbit correction"
+    if 'V' in plane and len(pvy) > 0 and len(pvysp) == 0:
+        print "WARNING: no VCOR for vertical orbit correction"
 
-    v0 = getOrbit(bpm)[:,0]
-    dk, resids, rank, s = np.linalg.lstsq(m, -1.0*v0)
-    eput(trim, dk)
-    
-    #v1 = getOrbit(bpm)[:,0]
-    #print np.shape(m), np.shape(v)
-
-    #import matplotlib.pylab as plt
-    #plt.clf()
-    #plt.plot(v0, '--')
-    #plt.plot(v1, '-x')
-    #plt.savefig('tmp.png')
-    #pass
+    if not machines._lat.orm:
+        print "ERROR: this lattice setting has no ORM data"
+    else:
+        correctOrbitPv(list(set(bpmpv)), list(set(trimpv)), machines._lat.orm)