+++ test/utest.py
@@ -7,6 +7,7 @@ import hla
 print "= HLA used:", hla.__path__
 
 #import utChanFinder
+import utElement
 import utLattice
 import utTwiss
 #import utAllPVs
@@ -17,12 +18,11 @@ loader = unittest.TestLoader()
 
 
 suite = unittest.TestSuite()
-#suite.addTests(loader.loadTestsFromModule(utLattice))
+suite.addTests(loader.loadTestsFromModule(utElement))
 suite.addTests(loader.loadTestsFromTestCase(utLattice.TestLatticeTxt))
 suite.addTests(loader.loadTestsFromTestCase(utLattice.TestLatticeCf))
-#suite.addTests(loader.loadTestsFromModule(utTwiss))
+suite.addTests(loader.loadTestsFromModule(utTwiss))
 
-#suite.addTests(loader.loadTestsFromModule(utLattice))
 #suite.addTests(loader.loadTestsFromModule(utAllPVs))
 #suite.addTests(loader.loadTestsFromModule(utOrbit))
 #suite.addTests(loader.loadTestsFromModule(utOrm))